Remove WEB_HMI form develop branch
Because `WEB_HMI` has been moved to the separate repository
`sdl_hmi`, it's unnecessary to hold it any more in `sdl_core` repo.

It been removed from `feature/extended_policy_functionality`, but
because it's not related to new policy and RC functionality it
will increase complexity of final review of this large features.

To reduce complexity of future pull request `WEB_HMI`
should be also removed from `develop` branch